Crate buddy_system_allocator

Modules

linked_list

Provide the intrusive LinkedList

Structs

FrameAllocator

A frame allocator that uses buddy system, requiring a global allocator

Heap

A heap that uses buddy system

LockedFrameAllocator

A locked version of FrameAllocator

LockedHeap

A locked version of Heap

LockedHeapWithRescue

A locked version of Heap with rescue before oom