buddy_system_allocator::linked_list

Struct LinkedList

pub struct LinkedList { /* private fields */ }

An intrusive linked list

A clean room implementation of the one used in CS140e 2018 Winter

Thanks Sergio Benitez for his excellent work, See CS140e for more information

Implementations

impl LinkedList

pub const fn new() -> LinkedList

Create a new LinkedList

pub fn is_empty(&self) -> bool

Return true if the list is empty

pub unsafe fn push(&mut self, item: *mut usize)

Push item to the front of the list

pub fn pop(&mut self) -> Option<*mut usize>

Try to remove the first item in the list

pub fn iter(&self) -> Iter<'_>

Return an iterator over the items in the list

pub fn iter_mut(&mut self) -> IterMut<'_>

Return an mutable iterator over the items in the list