buddy_system_allocator
Struct LockedHeapWithRescue

pub struct LockedHeapWithRescue { /* private fields */ }

A locked version of Heap with rescue before oom

Usage

Create a locked heap:

use buddy_system_allocator::*;
let heap = LockedHeapWithRescue::new(|heap: &mut Heap| {});

Before oom, the allocator will try to call rescue function and try for one more time.

Implementations

impl LockedHeapWithRescue

pub fn new(rescue: fn(_: &mut Heap)) -> LockedHeapWithRescue

Creates an empty heap